Move ros operators into separate file and respond to pull request comments.
